webapp/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
redis==5.0.1
elasticsearch==8.11.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
